requirements.txt
openai>=1.0
autogen-agentchat
autogen-ext[openai]
python-dotenv>=1.0
gradio>=4.0
